pyproject.toml
[tool.poetry]
name = "monotome"
version = "0.2.0"
description = "Declarative SQL migration system for modular and monorepo-based projects"
authors = ["Eric Gazoni <eric@adimian.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10, <4"
click = ">=8.1"
pyyaml = ">=6.0"
prettytable = ">=3.14"
sqlalchemy = ">=2"
jinja2 = ">=3.1"
sqlparse = ">=0.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
monotome = 'monotome.cli:main'
